+++ modules/eslint-plugin/spec/rules/prefix-selectors-with-select.spec.ts
@@ -183,6 +183,36 @@ export const selectCreateSelect = createSelectorFactory((projectionFun) =>
       ],
     }
   ),
+  // https://github.com/ngrx/platform/issues/3956
+  fromFixture(
+    `
+import {createFeatureSelector} from '@ngrx/store';
+
+export interface FileListResponseState {
+  loading: boolean;
+}
+
+const featureSelector = createFeatureSelector<FileListResponseState>("name");
+      ~~~~~~~~~~~~~~~ [${prefixSelectorsWithSelect} suggest]`,
+    {
+      suggestions: [
+        {
+          messageId: prefixSelectorsWithSelectSuggest,
+          data: {
+            name: 'selectFeatureSelector',
+          },
+          output: `
+import {createFeatureSelector} from '@ngrx/store';
+
+export interface FileListResponseState {
+  loading: boolean;
+}
+
+const selectFeatureSelector = createFeatureSelector<FileListResponseState>("name");`,
+        },
+      ],
+    }
+  ),
 ];
 
 ruleTester().run(path.parse(__filename).name, rule, {