+++ modules/eslint-plugin/src/utils/helper-functions/utils.ts
@@ -13,10 +13,11 @@ import {
   isProgram,
   isProperty,
   isPropertyDefinition,
-  isTemplateElement,
-  isTemplateLiteral,
   isTSTypeAnnotation,
   isTSTypeReference,
+  isTemplateElement,
+  isTemplateLiteral,
+  isTSInstantiationExpression,
 } from './guards';
 import { NGRX_MODULE_PATHS } from './ngrx-modules';
 
@@ -30,7 +31,8 @@ type InjectedParameter =
         | ConstructorFunctionExpression
         | (TSESTree.TSParameterProperty & {
             parent: ConstructorFunctionExpression;
-          });
+          })
+        | TSESTree.PropertyDefinition;
     };
 type InjectedParameterWithSourceCode = Readonly<{
   identifiers?: readonly InjectedParameter[];
@@ -315,6 +317,12 @@ function getInjectedParametersWithSourceCode(
   const { importSpecifier } =
     getImportDeclarationSpecifier(importDeclarations, importName) ?? {};
 
+  const injectImportDeclarations =
+    getImportDeclarations(sourceCode.ast, '@angular/core') ?? [];
+
+  const { importSpecifier: injectImportSpecifier } =
+    getImportDeclarationSpecifier(injectImportDeclarations, 'inject') ?? {};
+
   if (!importSpecifier) {
     return { sourceCode };
   }
@@ -324,8 +332,11 @@ function getInjectedParametersWithSourceCode(
   const identifiers = typedVariable?.references?.reduce<
     readonly InjectedParameter[]
   >((identifiers, { identifier: { parent } }) => {
+    if (!parent) {
+      return identifiers;
+    }
+
     if (
-      parent &&
       isTSTypeReference(parent) &&
       parent.parent &&
       isTSTypeAnnotation(parent.parent) &&
@@ -333,6 +344,23 @@ function getInjectedParametersWithSourceCode(
       isIdentifier(parent.parent.parent)
     ) {
       return identifiers.concat(parent.parent.parent as InjectedParameter);
+    }
+
+    const parentToCheck = isTSInstantiationExpression(parent)
+      ? parent.parent
+      : parent;
+
+    if (
+      parentToCheck &&
+      isCallExpression(parentToCheck) &&
+      isIdentifier(parentToCheck.callee) &&
+      parentToCheck.callee.name == 'inject' &&
+      parentToCheck.parent &&
+      isPropertyDefinition(parentToCheck.parent) &&
+      isIdentifier(parentToCheck.parent.key) &&
+      injectImportSpecifier
+    ) {
+      return identifiers.concat(parentToCheck.parent.key as InjectedParameter);
     }
 
     return identifiers;